WFLY-11821 Remove ha-singleton-service/with-backups quickstart since it's no longer supported without deprecated API usage
