+++ ha-singleton-service/pom.xml
@@ -31,9 +31,9 @@
     </parent>
 
     <artifactId>ha-singleton-service</artifactId>
-    <packaging>pom</packaging>
-    <name>Quickstart: HA Singleton Service (parent)</name>
-    <description>These quickstarts demonstrate two SingletonService usage patterns.</description>
+    <packaging>jar</packaging>
+    <name>Quickstart: HA Singleton Service</name>
+    <description>This quickstart demonstrates a singleton MSC service packaged in a jar.</description>
 
     <licenses>
         <license>
@@ -43,11 +43,6 @@
         </license>
     </licenses>
 
-    <modules>
-        <module>primary-only</module>
-        <module>with-backups</module>
-    </modules>
-    
     <dependencies>
         <dependency>
             <groupId>org.jboss.msc</groupId>
