+++ ha-singleton-service/src/main/java/org/jboss/as/quickstarts/ha/singleton/service/QueryingService.java
@@ -15,63 +15,63 @@
  * limitations under the License.
  */
 
-package org.jboss.as.quickstarts.ha.singleton.service.backups;
+package org.jboss.as.quickstarts.ha.singleton.service;
 
 import org.jboss.logging.Logger;
 import org.jboss.msc.service.Service;
+import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.StartContext;
 import org.jboss.msc.service.StartException;
 import org.jboss.msc.service.StopContext;
-import org.jboss.msc.value.InjectedValue;
-import org.wildfly.clustering.group.Group;
 import org.wildfly.clustering.group.Node;
 
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 /**
- * An implementation of service to demonstrate singleton service. After the service is started, it logs a message every
- * 5 seconds. This way we can illustrate which node is the primary and which nodes are the backups.
+ * This service will regularly query the singleton service and log where the singleton service is running. For the sake
+ * of illustration, the value returned by the singleton service {@link SingletonService#getValue()} is the node where it
+ * is currently running.
  *
  * @author Radoslav Husar
  */
-class SingletonService implements Service<Node> {
+class QueryingService implements Service<Void> {
 
     private Logger LOG = Logger.getLogger(this.getClass());
-
     private ScheduledExecutorService executor;
-    private InjectedValue<Group> group;
-    private boolean primary;
-
-    SingletonService(boolean primary, InjectedValue<Group> group) {
-        this.primary = primary;
-        this.group = group;
-    }
 
     @Override
     public void start(StartContext context) throws StartException {
-        LOG.infof("%s singleton service is starting.", this.serviceType());
+        LOG.info("Querying service is starting.");
 
         executor = Executors.newSingleThreadScheduledExecutor();
-        executor.scheduleAtFixedRate(
-                () -> LOG.infof("%s singleton service is running on node '%s'.", this.serviceType(), this.getValue()),
-                5, 5, TimeUnit.SECONDS);
+        executor.scheduleAtFixedRate(() -> {
+
+            @SuppressWarnings("unchecked")
+            ServiceController<Node> service = (ServiceController<Node>) context.getController().getServiceContainer()
+                    .getService(ServiceActivator.SINGLETON_SERVICE_NAME);
+            try {
+                Node node = service.awaitValue(5, TimeUnit.SECONDS);
+                LOG.infof("Singleton service is running on node '%s'.", node);
+            } catch (InterruptedException | TimeoutException | IllegalStateException e) {
+                LOG.warn("Failed to query singleton service.");
+            }
+
+        }, 5, 5, TimeUnit.SECONDS);
     }
 
     @Override
     public void stop(StopContext context) {
-        LOG.infof("%s singleton service is stopping.", this.serviceType());
+        LOG.info("Querying service is stopping.");
 
         executor.shutdown();
     }
 
     @Override
-    public Node getValue() throws IllegalStateException, IllegalArgumentException {
-        return this.group.getValue().getLocalNode();
+    public Void getValue() throws IllegalStateException, IllegalArgumentException {
+        return null;
     }
 
-    private String serviceType() {
-        return primary ? "Primary" : "Backup";
-    }
 }
