+++ ha-singleton-service/src/main/resources/META-INF/services/org.jboss.msc.service.ServiceActivator
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.jboss.as.quickstarts.ha.singleton.service.backups.ServiceActivator
+org.jboss.as.quickstarts.ha.singleton.service.ServiceActivator
